fix: check for default values when categorizing new properties
Properties with optional=false but with default values should be
categorized as informational, not required actions, since the
default will be used automatically.

Updated categorizer to check for nil Default field in addition to
Optional flag when determining if a property requires user action.

Added test case for required property with default value.